+++ tests/web_api/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import pkg_resources
-from openfisca_web_api.app import create_app
+from openfisca_web_api_preview.app import create_app
 
 TEST_COUNTRY_PACKAGE_NAME = 'openfisca_dummy_country'
 distribution = pkg_resources.get_distribution(TEST_COUNTRY_PACKAGE_NAME)
